requirements.txt
numpy==1.23.4
opencv-python==4.6.0.66
paddleocr==2.6.1.0
playsound==1.3.0
windows-capture-device-list==1.1.0